Move CropLatentsCoreInvocation to its own file. No functional changes.
